refactor: dom, event, utils 로 디렉터리를 나눔
화면에 출력하는 부분은 dom에 넣음
이벤트를 등록하는 부분은 event에 넣음
화면 조작 외의 모든 모듈은 utils에 넣음
