+++ src/Product.js
@@ -1,5 +1,5 @@
 import { STORAGE_KEY } from './constants.js';
-import * as storage from './storage.js';
+import * as storage from './utils/storage.js';
 
 export default class Product {
   constructor() {
