+++ src/dom/printChargeAmount.js
@@ -1,4 +1,4 @@
-import { DOM_ID_SELECTOR } from './constants.js';
+import { DOM_ID_SELECTOR } from '../constants.js';
 
 const printChargeAmount = (amount) => {
   const $chargeAmount = document.getElementById(DOM_ID_SELECTOR.chargeAmount);
