+++ src/utils/initializeContentComponent.js
@@ -1,16 +1,16 @@
-import { DOM_ID_SELECTOR, STORAGE_KEY } from './constants.js';
+import { DOM_ID_SELECTOR, STORAGE_KEY } from '../constants.js';
 import * as storage from './storage.js';
-import renderProductAddTemplate from './renderProductAddTemplate.js';
-import renderProductPurchaseTemplate from './renderProductPurchaseTemplate.js';
-import renderVendingMachineManageMenuTemplate from './renderVendingMachineManageMenuTemplate.js';
-import printProductManageTable from './printProductManageTable.js';
-import printChargeAmount from './printChargeAmount.js';
-import printProductPurchaseItemTable from './printProductPurchaseItemTable.js';
-import printVendingMachineAmountAndCoinTable from './printVendingMachineAmountAndCoinTable.js';
-import attachProductAddEvent from './attachProductAddEvent.js';
-import attachChargeEvent from './attachChargeEvent.js';
-import attachReturnCoinEvent from './attachReturnCoinEvent.js';
-import attachVendingMachineChargeEvent from './attachVendingMachineChargeEvent.js';
+import renderProductAddTemplate from '../dom/renderProductAddTemplate.js';
+import renderProductPurchaseTemplate from '../dom/renderProductPurchaseTemplate.js';
+import renderVendingMachineManageMenuTemplate from '../dom/renderVendingMachineManageMenuTemplate.js';
+import printProductManageTable from '../dom/printProductManageTable.js';
+import printChargeAmount from '../dom/printChargeAmount.js';
+import printProductPurchaseItemTable from '../dom/printProductPurchaseItemTable.js';
+import printVendingMachineAmountAndCoinTable from '../dom/printVendingMachineAmountAndCoinTable.js';
+import attachProductAddEvent from '../event/attachProductAddEvent.js';
+import attachChargeEvent from '../event/attachChargeEvent.js';
+import attachReturnCoinEvent from '../event/attachReturnCoinEvent.js';
+import attachVendingMachineChargeEvent from '../event/attachVendingMachineChargeEvent.js';
 
 const initializeProductAddComponent = (vendingMachine) => {
   renderProductAddTemplate();
